+++ TechTalk.SpecFlow/Events/IExecutionEventListener.cs
@@ -1,0 +1,7 @@
+﻿namespace TechTalk.SpecFlow.Events
+{
+    public interface IExecutionEventListener
+    {
+        void OnEvent(IExecutionEvent executionEvent);
+    }
+}
